+++ knowthenics-data/src/main/java/ru/dpopkov/knowthenics/model/Category.java
@@ -4,6 +4,8 @@ import lombok.*;
 
 import javax.persistence.Entity;
 import javax.persistence.Table;
+import javax.validation.constraints.NotBlank;
+import javax.validation.constraints.Size;
 
 @Getter
 @Setter
@@ -12,7 +14,10 @@ import javax.persistence.Table;
 @Entity
 @Table(name = "categories")
 public class Category extends BaseEntity {
+    @NotBlank
+    @Size(min = 3, max = 255)
     private String name;
+    @Size(max = 255)
     private String description;
 
     @Builder