+++ knowthenics-data/src/main/java/ru/dpopkov/knowthenics/model/Source.java
@@ -1,11 +1,14 @@
 package ru.dpopkov.knowthenics.model;
 
 import lombok.*;
+import org.hibernate.validator.constraints.URL;
 
 import javax.persistence.Entity;
 import javax.persistence.EnumType;
 import javax.persistence.Enumerated;
 import javax.persistence.Table;
+import javax.validation.constraints.NotBlank;
+import javax.validation.constraints.Size;
 
 @Getter
 @Setter
@@ -14,11 +17,17 @@ import javax.persistence.Table;
 @Entity
 @Table(name = "sources")
 public class Source extends BaseEntity {
+    @Size(min = 3, max = 128)
+    @NotBlank
     private String shortTitle;
+    @Size(max = 255)
     private String fullTitle;
+    @URL
+    @NotBlank
     private String url;
     @Enumerated(EnumType.STRING)
     private SourceType sourceType;
+    @Size(max = 255)
     private String description;
 
     @Builder