+++ History.md
@@ -2,6 +2,7 @@
 
 #### 2021-01-24
 * Add validation to Answer and AnswerController. Closes #46.
+* Add validation to Category, KeyTerm, and Source. Closes #47.
 
 #### 2021-01-23
 * Add validation to Question and QuestionController.